+++ .dotnet/examples/Embeddings/Example01_SimpleEmbedding.cs
@@ -16,7 +16,7 @@ public partial class EmbeddingExamples
             + " attractions. We highly recommend this hotel.";
 
         Embedding embedding = client.GenerateEmbedding(description);
-        ReadOnlyMemory<float> vector = embedding.GetAsFloat();
+        ReadOnlyMemory<float> vector = embedding.ToFloats();
 
         Console.WriteLine($"Dimension: {vector.Length}");
         Console.WriteLine($"Floats: ");